Hooked up video from client to hub
Clients call getDisplayMedia (and then the user has to pick that tab because the API requires user interaction) and then share the resulting MediaStream with the RTC channel.

The hub receives that stream and jams it into a video element.

This currently requires that the user click "play" on the video element. Not sure why.